Register event-listener through 'Contract' object conditionally
