+++ src/main/java/hlf/java/rest/client/listener/ChaincodeEventListener.java
@@ -4,7 +4,10 @@ import hlf.java.rest.client.model.EventType;
 import hlf.java.rest.client.service.EventPublishService;
 import hlf.java.rest.client.util.FabricClientConstants;
 import hlf.java.rest.client.util.FabricEventParseUtil;
+import java.nio.charset.StandardCharsets;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.hyperledger.fabric.gateway.ContractEvent;
 import org.hyperledger.fabric.sdk.BlockInfo;
 import org.hyperledger.fabric.sdk.ChaincodeEvent;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,23 +19,55 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ConditionalOnProperty(prefix = "fabric.events", name = "enable", havingValue = "true")
 @RefreshScope
-public class ChaincodeEventListener {
+public class CustomChaincodeEventListener {
 
   @Autowired(required = false)
   private EventPublishService eventPublishService;
 
   private static String eventTxnId = FabricClientConstants.FABRIC_TRANSACTION_ID;
 
+  public void chaincodeEventListener(ContractEvent contractEvent) {
+
+    String chaincodeId = contractEvent.getChaincodeId();
+    String eventName = contractEvent.getName();
+    String txId = contractEvent.getTransactionEvent().getTransactionID();
+    String channelName = contractEvent.getTransactionEvent().getChannelId();
+    long blockNumber = contractEvent.getTransactionEvent().getBlockEvent().getBlockNumber();
+    String payload =
+        contractEvent.getPayload().isPresent()
+            ? new String(contractEvent.getPayload().get(), StandardCharsets.UTF_8)
+            : StringUtils.EMPTY;
+
+    publishChaincodeEvent(txId, chaincodeId, eventName, payload, channelName, blockNumber);
+  }
+
+  @Deprecated
   public void listener(
       String handle, BlockInfo blockEvent, ChaincodeEvent chaincodeEvent, String channelName) {
 
-    synchronized (this) {
-      if (!chaincodeEvent.getTxId().equalsIgnoreCase(eventTxnId)) {
+    long blockNumber = blockEvent.getBlockNumber();
+    String txId = chaincodeEvent.getTxId();
+    String chaincodeId = chaincodeEvent.getChaincodeId();
+    String eventName = chaincodeEvent.getEventName();
+    String payload = new String(chaincodeEvent.getPayload(), StandardCharsets.UTF_8);
 
-        log.info("Chaincode ID: {}", chaincodeEvent.getChaincodeId());
-        log.info("Event Name: {}", chaincodeEvent.getEventName());
-        log.info("Transaction ID: {}", chaincodeEvent.getTxId());
-        log.info("Payload: {}", new String(chaincodeEvent.getPayload()));
+    publishChaincodeEvent(txId, chaincodeId, eventName, payload, channelName, blockNumber);
+  }
+
+  private void publishChaincodeEvent(
+      String txId,
+      String chaincodeId,
+      String eventName,
+      String payload,
+      String channelName,
+      long blockNumber) {
+    synchronized (this) {
+      if (!txId.equalsIgnoreCase(eventTxnId)) {
+
+        log.info("Chaincode ID: {}", chaincodeId);
+        log.info("Event Name: {}", eventName);
+        log.info("Transaction ID: {}", txId);
+        log.info("Payload: {}", payload);
         log.info("Channel Name: {}", channelName);
 
         if (eventPublishService == null) {
@@ -42,18 +77,14 @@ public class ChaincodeEventListener {
 
         eventPublishService.publishChaincodeEvents(
             FabricEventParseUtil.createEventStructure(
-                new String(chaincodeEvent.getPayload()),
-                "",
-                chaincodeEvent.getTxId(),
-                blockEvent.getBlockNumber(),
-                EventType.CHAINCODE_EVENT),
-            chaincodeEvent.getChaincodeId(),
-            chaincodeEvent.getTxId(),
-            chaincodeEvent.getEventName(),
+                payload, "", txId, blockNumber, EventType.CHAINCODE_EVENT),
+            chaincodeId,
+            txId,
+            eventName,
             channelName);
-        eventTxnId = chaincodeEvent.getTxId();
+        eventTxnId = txId;
       } else {
-        log.debug("Duplicate Transaction; ID: {}", chaincodeEvent.getTxId());
+        log.debug("Duplicate Transaction; ID: {}", txId);
       }
     }
   }
